Create domains for applications and students/alternate-emails
